pyproject.toml
[project]
name = "cpa-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "flask",
  "python-dotenv",
  "google-genai",
]
